Add twoSum.cpp as solution to LeetCode 1
A simple solution to LeetCode 1 that runs in O(n^2) time where the
indices of the arrays that add up to a target is found by iterating
through number pairs until a pair is found that is equal to the target
value. The index pair is saved in a vector and is returned to solve the
problem.