Add solution to adding 2 singly linked lists
A solution to adding 2 singly linked lists where the elements are added
by adding the contents of a digit from both linked lists if they are
non-zero to a carry variable that initializes the result as the addition
of the 2 digits % 10. This results in a correct current digit for
discarding numbers that would be larger than one digit and the carry
variable is divided by 10 for future iterations so that the carry can
be added to the next result correctly.